+++ src/connectedk8s/azext_connectedk8s/custom.py
@@ -172,10 +172,10 @@ def create_connectedk8s(cmd, client, resource_group_name, cluster_name, correlat
         raise ManualInterrupt('Process terminated externally.')
 
     # Pre onboarding checks
+    diagnostic_checks = "Failed"
     try:
         # if aks_hci skip, otherwise continue to perform pre-onboarding check
         if not aks_hci:
-            diagnostic_checks = "Failed"
             batchv1_api_instance = kube_client.BatchV1Api()
             storage_space_available = True